Moved 'weekday' into a _common.py module.
